product_service/product_service/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
gunicorn==20.1.0
beanie==1.17
pydantic==1.10.5
starlette==0.25.0
passlib[bcrypt]
python-jose
python-dotenv==1.0.0
Jinja2
